Remove burn.cpp.
