+++ ksp_plugin/ksp_plugin.vcxproj
@@ -41,7 +41,6 @@
     <ClCompile Include="..\journal\profiles.cpp" />
     <ClCompile Include="..\journal\recorder.cpp" />
     <ClCompile Include="..\numerics\cbrt.cpp" />
-    <ClCompile Include="burn.cpp" />
     <ClCompile Include="celestial.cpp" />
     <ClCompile Include="flight_plan.cpp" />
     <ClCompile Include="identification.cpp" />
