+++ ksp_plugin/ksp_plugin.vcxproj.filters
@@ -88,9 +88,6 @@
     <ClCompile Include="flight_plan.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="burn.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="interface_flight_plan.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
