+++ ksp_plugin_test/ksp_plugin_test.vcxproj
@@ -11,7 +11,6 @@
     <ClCompile Include="..\base\version.generated.cc" />
     <ClCompile Include="..\journal\profiles.cpp" />
     <ClCompile Include="..\journal\recorder.cpp" />
-    <ClCompile Include="..\ksp_plugin\burn.cpp" />
     <ClCompile Include="..\ksp_plugin\celestial.cpp" />
     <ClCompile Include="..\ksp_plugin\flight_plan.cpp" />
     <ClCompile Include="..\ksp_plugin\identification.cpp" />
